+++ common/src/main/java/com/st0x0ef/stellaris/common/energy/impl/SimpleEnergyContainer.java
@@ -38,7 +38,7 @@ public class SimpleEnergyContainer implements EnergyContainer {
 
     @Override
     public long insertEnergy(long maxAmount, boolean simulate) {
-        long inserted = (long) Mth.clamp(maxAmount, 0, Math.min(maxInsert(), getMaxCapacity() - getStoredEnergy()));
+        long inserted = Mth.clamp(maxAmount, 0, Math.min(maxInsert(), getMaxCapacity() - getStoredEnergy()));
         if (simulate) return inserted;
         this.setEnergy(this.energy + inserted);
         return inserted;
@@ -46,7 +46,7 @@ public class SimpleEnergyContainer implements EnergyContainer {
 
     @Override
     public long extractEnergy(long maxAmount, boolean simulate) {
-        long extracted = (long) Mth.clamp(maxAmount, 0, Math.min(maxExtract(), getStoredEnergy()));
+        long extracted = Mth.clamp(maxAmount, 0, Math.min(maxExtract(), getStoredEnergy()));
         if (simulate) return extracted;
         this.setEnergy(this.energy - extracted);
         return extracted;
@@ -54,7 +54,7 @@ public class SimpleEnergyContainer implements EnergyContainer {
 
     @Override
     public long internalInsert(long maxAmount, boolean simulate) {
-        long inserted = (long) Mth.clamp(maxAmount, 0, getMaxCapacity() - getStoredEnergy());
+        long inserted = Mth.clamp(maxAmount, 0, getMaxCapacity() - getStoredEnergy());
         if (simulate) return inserted;
         this.setEnergy(this.energy + inserted);
         return inserted;
@@ -62,7 +62,7 @@ public class SimpleEnergyContainer implements EnergyContainer {
 
     @Override
     public long internalExtract(long maxAmount, boolean simulate) {
-        long extracted = (long) Mth.clamp(maxAmount, 0, getStoredEnergy());
+        long extracted = Mth.clamp(maxAmount, 0, getStoredEnergy());
         if (simulate) return extracted;
         this.setEnergy(this.energy - extracted);
         return extracted;
